Rename domain file namespace
