+++ src/CaptureTool.Application.Abstractions/Features/AudioEdit/OpenAudioEditPage/OpenAudioEditPageRequest.cs
@@ -1,4 +1,4 @@
-using CaptureTool.Domain.Files;
+using CaptureTool.Domain.FileSystem;
 
 namespace CaptureTool.Application.Abstractions.Features.AudioEdit.OpenAudioEditPage;
