+++ src/CaptureTool.Application.Abstractions/Features/ImageEdit/ChromaKey/IChromaKeyService.cs
@@ -1,4 +1,4 @@
-using CaptureTool.Domain.Files;
+using CaptureTool.Domain.FileSystem;
 using System.Drawing;
 
 namespace CaptureTool.Application.Abstractions.Features.ImageEdit.ChromaKey;
