+++ src/CaptureTool.Infrastructure.Capture.Windows/WindowsAudioRecorder.cs
@@ -1,6 +1,6 @@
 using CaptureKit.Abstractions;
 using CaptureTool.Domain.Capture;
-using CaptureTool.Domain.Files;
+using CaptureTool.Domain.FileSystem;
 
 namespace CaptureTool.Infrastructure.Capture.Windows;
