+++ src/CaptureTool.Infrastructure.Tests/Settings/LocalSettingsServiceTests.cs
@@ -2,7 +2,7 @@ using CaptureTool.Application.Abstractions.Logging;
 using CaptureTool.Application.Abstractions.Settings;
 using CaptureTool.Application.Abstractions.Storage;
 using CaptureTool.Application.Features.Settings.Definitions;
-using CaptureTool.Domain.Files;
+using CaptureTool.Domain.FileSystem;
 using CaptureTool.Infrastructure.Settings;
 using System.Text.Json.Serialization.Metadata;
