+++ src/CaptureTool.Presentation.Tests/Features/ImageEditPageViewModelShareTests.cs
@@ -8,7 +8,7 @@ using CaptureTool.Application.Abstractions.Share;
 using CaptureTool.Application.Abstractions.Storage;
 using CaptureTool.Application.Abstractions.Store;
 using CaptureTool.Domain.Edit.Drawable;
-using CaptureTool.Domain.Files;
+using CaptureTool.Domain.FileSystem;
 using CaptureTool.Presentation.Features.ImageEdit;
 using Moq;
 using System.Drawing;
