+++ src/CaptureTool.Presentation.Tests/Features/VideoEditPageViewModelTrimTests.cs
@@ -5,7 +5,7 @@ using CaptureTool.Application.Abstractions.Media;
 using CaptureTool.Application.Abstractions.Storage;
 using CaptureTool.Application.Abstractions.UseCases;
 using CaptureTool.Domain.Capture;
-using CaptureTool.Domain.Files;
+using CaptureTool.Domain.FileSystem;
 using CaptureTool.Presentation.Features.VideoEdit;
 using Moq;
