+++ src/CaptureTool.Presentation.Windows.WinUI/Xaml/Windows/SelectionOverlayHost.cs
@@ -1,6 +1,6 @@
 using CaptureTool.Application.Abstractions.Features.Navigation;
 using CaptureTool.Domain.Capture;
-using CaptureTool.Domain.Files;
+using CaptureTool.Domain.FileSystem;
 using CaptureTool.Infrastructure.Capture.Windows;
 using CaptureTool.Presentation.Features.SelectionOverlay;
 using CaptureTool.Presentation.Windows.WinUI.Utils;
